Shrink the images.
- Delete cached files used for installing perl modules.  (Saves 650MB!)
- Bookworm only: base on the slim-bookworm base image.

bookworm before: 1.6GB
slim-bookworm: 1.03GB
slim-bookworm clean: 572MB

The larger image (with a ton of files) takes a long time to download and
initialize.  (40% of the time to run pause tests in GitHub actions.) A
smaller image is faster to download and unpack.

There is a tiny chance that someone relying on the new bookworm image is
relying on a package that isn't installed by default in the slim images, but
the bookworm image has only existed for a week or so, so there are likely to
be very few users.